pyproject.toml
[tool.poetry]
name = "digesto-hybrid-search"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
simple-term-menu = "^1.6.4"
tqdm = "^4.66.5"
sentence-transformers = "^3.0.1"
pymupdf = "^1.24.9"
nltk = "3.8.1" # Updating it causes an error with Punkt
python-terrier = "^0.10.1"
numpy = "1.26.4"
logging = "^0.4.9.6"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

